moved enum entity to specific file
